Generic einsum loop now uses a sum_of_products function

+++ src/macros.rs
@@ -114,7 +114,7 @@ macro_rules! test_for_all_dtypes {
 #[macro_export]
 macro_rules! assert_almost_eq {
     ($left:expr, $right:expr) => {
-        if ((&($left) - &($right)).max().flatiter().next().unwrap() as f32).abs() > 0.2 {
+        if ((&($left) - &($right)).max().flatiter().next().unwrap() as f64).abs() > 0.2 {
             assert_eq!($left, $right);
         }
     };
